requirements.txt
Flask
tensorflow-cpu
numpy
gunicorn 
requests 